requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
openai
tiktoken
faiss-cpu
python-dotenv

